+++ contrib/platform/src/com/sun/jna/platform/WindowUtils.java
@@ -46,6 +46,8 @@ import java.awt.geom.PathIterator;
 import java.awt.image.BufferedImage;
 import java.awt.image.Raster;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.LinkedList;
 import java.util.List;
 
 import javax.swing.Icon;
@@ -70,18 +72,29 @@ import com.sun.jna.platform.unix.X11.XVisualInfo;
 import com.sun.jna.platform.unix.X11.Xext;
 import com.sun.jna.platform.unix.X11.Xrender.XRenderPictFormat;
 import com.sun.jna.platform.win32.GDI32;
+import com.sun.jna.platform.win32.Kernel32;
+import com.sun.jna.platform.win32.Psapi;
 import com.sun.jna.platform.win32.User32;
+import com.sun.jna.platform.win32.Win32Exception;
+import com.sun.jna.platform.win32.WinNT;
+import com.sun.jna.platform.win32.WinDef.DWORDByReference;
 import com.sun.jna.platform.win32.WinDef.HBITMAP;
 import com.sun.jna.platform.win32.WinDef.HDC;
+import com.sun.jna.platform.win32.WinDef.HICON;
 import com.sun.jna.platform.win32.WinDef.HRGN;
 import com.sun.jna.platform.win32.WinDef.HWND;
+import com.sun.jna.platform.win32.WinDef.POINT;
+import com.sun.jna.platform.win32.WinDef.RECT;
 import com.sun.jna.platform.win32.WinGDI;
+import com.sun.jna.platform.win32.WinGDI.BITMAP;
 import com.sun.jna.platform.win32.WinGDI.BITMAPINFO;
+import com.sun.jna.platform.win32.WinGDI.BITMAPINFOHEADER;
+import com.sun.jna.platform.win32.WinGDI.ICONINFO;
 import com.sun.jna.platform.win32.WinNT.HANDLE;
 import com.sun.jna.platform.win32.WinUser;
 import com.sun.jna.platform.win32.WinUser.BLENDFUNCTION;
-import com.sun.jna.platform.win32.WinDef.POINT;
 import com.sun.jna.platform.win32.WinUser.SIZE;
+import com.sun.jna.platform.win32.WinUser.WNDENUMPROC;
 import com.sun.jna.ptr.ByteByReference;
 import com.sun.jna.ptr.IntByReference;
 import com.sun.jna.ptr.PointerByReference;
@@ -123,6 +136,8 @@ import com.sun.jna.ptr.PointerByReference;
  * component, causing creation of the heavyweight peer, changing this
  * property has no effect. 
  * @see <a href="http://developer.apple.com/technotes/tn2007/tn2196.html#APPLE_AWT_DRAGGABLEWINDOWBACKGROUND">Apple Technote 2007</a>
+ * 
+ * @author Andreas "PAX" L&uuml;ck, onkelpax-git[at]yahoo.de
  */
 // TODO: setWindowMask() should accept a threshold; some cases want a
 // 50% threshold, some might want zero/non-zero
@@ -556,6 +571,115 @@ public class WindowUtils {
                 }
             }
         }
+
+		/**
+		 * Obtains the set icon for the window associated with the specified
+		 * window handle.
+		 *
+		 * @param hwnd
+		 *            The concerning window handle.
+		 * @return Either the window's icon or {@code null} if an error
+		 *         occurred.
+		 * 
+		 * @throws UnsupportedOperationException
+		 *             Thrown if this method wasn't yet implemented for the
+		 *             current platform.
+		 */
+		protected BufferedImage getWindowIcon(final HWND hwnd) {
+			throw new UnsupportedOperationException(
+					"This platform is not supported, yet.");
+		}
+
+		/**
+		 * Detects the size of an icon.
+		 *
+		 * @param hIcon
+		 *            The icon handle type.
+		 * @return Either the requested icon's dimension or an {@link Dimension}
+		 *         instance of {@code (0, 0)}.
+		 * 
+		 * @throws UnsupportedOperationException
+		 *             Thrown if this method wasn't yet implemented for the
+		 *             current platform.
+		 */
+		protected Dimension getIconSize(final HICON hIcon) {
+			throw new UnsupportedOperationException(
+					"This platform is not supported, yet.");
+		}
+
+		/**
+		 * Requests a list of all currently available Desktop windows.
+		 *
+		 * @param onlyVisibleWindows
+		 *            Specifies whether only currently visible windows will be
+		 *            considered ({@code true}). That are windows which are not
+		 *            minimized. The {@code WS_VISIBLE} flag will be checked
+		 *            (see: <a href=
+		 *            "https://msdn.microsoft.com/de-de/library/windows/desktop/ms633530%28v=vs.85%29.aspx"
+		 *            >User32.IsWindowVisible(HWND)</a>).
+		 *
+		 * @return A list with all windows and some detailed information.
+		 * 
+		 * @throws UnsupportedOperationException
+		 *             Thrown if this method wasn't yet implemented for the
+		 *             current platform.
+		 */
+		protected List<DesktopWindow> getAllWindows(
+				final boolean onlyVisibleWindows) {
+			throw new UnsupportedOperationException(
+					"This platform is not supported, yet.");
+		}
+
+		/**
+		 * Tries to obtain the Window's title which belongs to the specified
+		 * window handle.
+		 *
+		 * @param hwnd
+		 *            The concerning window handle.
+		 * @return Either the title or an empty string of no title was found or
+		 *         an error occurred.
+		 * 
+		 * @throws UnsupportedOperationException
+		 *             Thrown if this method wasn't yet implemented for the
+		 */
+		protected String getWindowTitle(final HWND hwnd) {
+			throw new UnsupportedOperationException(
+					"This platform is not supported, yet.");
+		}
+
+	    /**
+	     * Detects the full file path of the process associated with the specified
+	     * window handle.
+	     *
+	     * @param hwnd
+	     *            The concerning window handle for which the PE file path is
+	     *            required.
+	     * @return The full file path of the PE file that is associated with the
+	     *         specified window handle.
+		 * 
+		 * @throws UnsupportedOperationException
+		 *             Thrown if this method wasn't yet implemented for the
+	     */
+	    protected  String getProcessFilePath(final HWND hwnd){
+	    	throw new UnsupportedOperationException(
+					"This platform is not supported, yet.");
+	    }
+
+		/**
+		 * Requests the location and size of the window associated with the
+		 * specified window handle.
+		 *
+		 * @param hwnd
+		 *            The concerning window handle.
+		 * @return The location and size of the window.
+		 * 
+		 * @throws UnsupportedOperationException
+		 *             Thrown if this method wasn't yet implemented for the
+		 */
+		protected Rectangle getWindowLocationAndSize(final HWND hwnd) {
+			throw new UnsupportedOperationException(
+					"This platform is not supported, yet.");
+		}
     }
     /** Canonical lazy loading of a singleton. */
     private static class Holder {
@@ -932,7 +1056,206 @@ public class WindowUtils {
             }
             setWindowRegion(w, region);
         }
-    }
+
+		@Override
+		public BufferedImage getWindowIcon(final HWND hwnd) {
+			// request different kind of icons if any solution fails
+			final DWORDByReference hIconNumber = new DWORDByReference();
+			long result = User32.INSTANCE.SendMessageTimeout(hwnd,
+					WinUser.WM_GETICON, WinUser.ICON_BIG, 0,
+					WinUser.SMTO_ABORTIFHUNG, 500, hIconNumber);
+			if (result == 0)
+				result = User32.INSTANCE.SendMessageTimeout(hwnd,
+						WinUser.WM_GETICON, WinUser.ICON_SMALL, 0,
+						WinUser.SMTO_ABORTIFHUNG, 500, hIconNumber);
+			if (result == 0)
+				result = User32.INSTANCE.SendMessageTimeout(hwnd,
+						WinUser.WM_GETICON, WinUser.ICON_SMALL2, 0,
+						WinUser.SMTO_ABORTIFHUNG, 500, hIconNumber);
+			if (result == 0) {
+				result = User32.INSTANCE.GetClassLongPtr(hwnd,
+						WinUser.GCLP_HICON);
+				hIconNumber.getValue().setValue(result);
+			}
+			if (result == 0) {
+				result = User32.INSTANCE.GetClassLongPtr(hwnd,
+						WinUser.GCLP_HICONSM);
+				hIconNumber.getValue().setValue(result);
+			}
+			if (result == 0)
+				return null;
+
+			// draw native icon into Java image
+			final HICON hIcon = new HICON(new Pointer(hIconNumber.getValue()
+					.longValue()));
+			final Dimension iconSize = getIconSize(hIcon);
+			if (iconSize.width == 0 || iconSize.height == 0)
+				return null;
+
+			final int width = iconSize.width;
+			final int height = iconSize.height;
+			final short depth = 24;
+
+			final byte[] lpBitsColor = new byte[width * height * depth / 8];
+			final Pointer lpBitsColorPtr = new Memory(lpBitsColor.length);
+			final byte[] lpBitsMask = new byte[width * height * depth / 8];
+			final Pointer lpBitsMaskPtr = new Memory(lpBitsMask.length);
+			final BITMAPINFO bitmapInfo = new BITMAPINFO();
+			final BITMAPINFOHEADER hdr = new BITMAPINFOHEADER();
+			
+			bitmapInfo.bmiHeader = hdr;
+			hdr.biWidth = width;
+			hdr.biHeight = height;
+			hdr.biPlanes = 1;
+			hdr.biBitCount = depth;
+			hdr.biCompression = 0;
+			hdr.write();
+			bitmapInfo.write();
+
+			final HDC hDC = User32.INSTANCE.GetDC(null);
+			final ICONINFO iconInfo = new ICONINFO();
+			User32.INSTANCE.GetIconInfo(hIcon, iconInfo);
+			iconInfo.read();
+			GDI32.INSTANCE.GetDIBits(hDC, iconInfo.hbmColor, 0, height,
+					lpBitsColorPtr, bitmapInfo, 0);
+			lpBitsColorPtr.read(0, lpBitsColor, 0, lpBitsColor.length);
+			GDI32.INSTANCE.GetDIBits(hDC, iconInfo.hbmMask, 0, height,
+					lpBitsMaskPtr, bitmapInfo, 0);
+			lpBitsMaskPtr.read(0, lpBitsMask, 0, lpBitsMask.length);
+			final BufferedImage image = new BufferedImage(width, height,
+					BufferedImage.TYPE_INT_ARGB);
+
+			int r, g, b, a, argb;
+			int x = 0, y = height - 1;
+			for (int i = 0; i < lpBitsColor.length; i = i + 3) {
+				b = lpBitsColor[i] & 0xFF;
+				g = lpBitsColor[i + 1] & 0xFF;
+				r = lpBitsColor[i + 2] & 0xFF;
+				a = 0xFF - lpBitsMask[i] & 0xFF;
+				argb = (a << 24) | (r << 16) | (g << 8) | b;
+				image.setRGB(x, y, argb);
+				x = (x + 1) % width;
+				if (x == 0)
+					y--;
+			}
+
+			User32.INSTANCE.ReleaseDC(null, hDC);
+
+			return image;
+		}
+
+		@Override
+		public Dimension getIconSize(final HICON hIcon) {
+			final ICONINFO iconInfo = new ICONINFO();
+			try {
+				if (!User32.INSTANCE.GetIconInfo(hIcon, iconInfo))
+					return new Dimension();
+				iconInfo.read();
+
+				final BITMAP bmp = new BITMAP();
+				if (iconInfo.hbmColor != null
+						&& iconInfo.hbmColor.getPointer() != Pointer.NULL) {
+					final int nWrittenBytes = GDI32.INSTANCE.GetObject(
+							iconInfo.hbmColor, bmp.size(), bmp.getPointer());
+					bmp.read();
+					if (nWrittenBytes > 0)
+						return new Dimension(bmp.bmWidth.intValue(),
+								bmp.bmHeight.intValue());
+				} else if (iconInfo.hbmMask != null
+						&& iconInfo.hbmMask.getPointer() != Pointer.NULL) {
+					final int nWrittenBytes = GDI32.INSTANCE.GetObject(
+							iconInfo.hbmMask, bmp.size(), bmp.getPointer());
+					bmp.read();
+					if (nWrittenBytes > 0)
+						return new Dimension(bmp.bmWidth.intValue(), bmp.bmHeight.intValue() / 2);
+				}
+			} finally {
+				if (iconInfo.hbmColor != null
+						&& iconInfo.hbmColor.getPointer() != Pointer.NULL)
+					GDI32.INSTANCE.DeleteObject(iconInfo.hbmColor);
+				if (iconInfo.hbmMask != null
+						&& iconInfo.hbmMask.getPointer() != Pointer.NULL)
+					GDI32.INSTANCE.DeleteObject(iconInfo.hbmMask);
+			}
+
+			return new Dimension();
+		}
+
+		@Override
+		public List<DesktopWindow> getAllWindows(
+				final boolean onlyVisibleWindows) {
+			final List<DesktopWindow> result = new LinkedList<DesktopWindow>();
+
+			final WNDENUMPROC lpEnumFunc = new WNDENUMPROC() {
+				@Override
+				public boolean callback(final HWND hwnd, final Pointer arg1) {
+					try {
+						final boolean visible = !onlyVisibleWindows
+								|| User32.INSTANCE.IsWindowVisible(hwnd);
+						if (visible) {
+							final String title = getWindowTitle(hwnd);
+							final String filePath = getProcessFilePath(hwnd);
+							final Rectangle locAndSize = getWindowLocationAndSize(hwnd);
+							result.add(new DesktopWindow(hwnd, title, filePath,
+									locAndSize));
+						}
+					} catch (final Exception e) {
+						e.printStackTrace();
+					}
+
+					return true;
+				}
+			};
+
+			if (!User32.INSTANCE.EnumWindows(lpEnumFunc, null))
+				throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+
+			return result;
+		}
+
+		@Override
+		public String getWindowTitle(final HWND hwnd) {
+			final int requiredLength = User32.INSTANCE
+					.GetWindowTextLength(hwnd) + 1;
+			final char[] title = new char[requiredLength];
+			final int length = User32.INSTANCE.GetWindowText(hwnd, title,
+					title.length);
+
+			return Native.toString(Arrays.copyOfRange(title, 0, length));
+		}
+
+		@Override
+		public String getProcessFilePath(final HWND hwnd) {
+			final char[] filePath = new char[1025];
+			final IntByReference pid = new IntByReference();
+			User32.INSTANCE.GetWindowThreadProcessId(hwnd, pid);
+
+			final HANDLE process = Kernel32.INSTANCE.OpenProcess(
+					WinNT.PROCESS_QUERY_INFORMATION | WinNT.PROCESS_VM_READ,
+					false, pid.getValue());
+			if (process == null
+					&& Kernel32.INSTANCE.GetLastError() != WinNT.ERROR_ACCESS_DENIED)
+				throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+
+			final int length = Psapi.INSTANCE.GetModuleFileNameExW(process,
+					null, filePath, filePath.length);
+			if (length == 0
+					&& Kernel32.INSTANCE.GetLastError() != WinNT.ERROR_INVALID_HANDLE)
+				throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+
+			return Native.toString(filePath).trim();
+		}
+
+		@Override
+		public Rectangle getWindowLocationAndSize(final HWND hwnd) {
+			final RECT lpRect = new RECT();
+			if (!User32.INSTANCE.GetWindowRect(hwnd, lpRect))
+				throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+
+			return new Rectangle(lpRect.left, lpRect.top, Math.abs(lpRect.right
+					- lpRect.left), Math.abs(lpRect.bottom - lpRect.top));
+		}
+	}
 
     private static class MacWindowUtils extends NativeWindowUtils {
         public boolean isWindowAlphaSupported() {
@@ -1548,4 +1871,86 @@ public class WindowUtils {
     public static void setWindowTransparent(Window w, boolean transparent) {
         getInstance().setWindowTransparent(w, transparent);
     }
+
+	/**
+	 * Obtains the set icon for the window associated with the specified
+	 * window handle.
+	 *
+	 * @param hwnd
+	 *            The concerning window handle.
+	 * @return Either the window's icon or {@code null} if an error
+	 *         occurred.
+	 */
+	public  static BufferedImage getWindowIcon(final HWND hwnd) {
+		return getInstance().getWindowIcon(hwnd);
+	}
+
+	/**
+	 * Detects the size of an icon.
+	 *
+	 * @param hIcon
+	 *            The icon handle type.
+	 * @return Either the requested icon's dimension or an {@link Dimension}
+	 *         instance of {@code (0, 0)}.
+	 */
+	public static Dimension getIconSize(final HICON hIcon) {
+		return getInstance().getIconSize(hIcon);
+	}
+
+	/**
+	 * Requests a list of all currently available Desktop windows.
+	 *
+	 * @param onlyVisibleWindows
+	 *            Specifies whether only currently visible windows will be
+	 *            considered ({@code true}). That are windows which are not
+	 *            minimized. The {@code WS_VISIBLE} flag will be checked (see:
+	 *            <a href=
+	 *            "https://msdn.microsoft.com/de-de/library/windows/desktop/ms633530%28v=vs.85%29.aspx"
+	 *            >User32.IsWindowVisible(HWND)</a>).
+	 *
+	 * @return A list with all windows and some detailed information.
+	 */
+	public static List<DesktopWindow> getAllWindows(
+			final boolean onlyVisibleWindows) {
+		return getInstance().getAllWindows(onlyVisibleWindows);
+	}
+
+	/**
+	 * Tries to obtain the Window's title which belongs to the specified window
+	 * handle.
+	 *
+	 * @param hwnd
+	 *            The concerning window handle.
+	 * @return Either the title or an empty string of no title was found or an
+	 *         error occurred.
+	 */
+	public static String getWindowTitle(final HWND hwnd) {
+		return getInstance().getWindowTitle(hwnd);
+	}
+
+	/**
+	 * Detects the full file path of the process associated with the specified
+	 * window handle.
+	 *
+	 * @param hwnd
+	 *            The concerning window handle for which the PE file path is
+	 *            required.
+	 * @return The full file path of the PE file that is associated with the
+	 *         specified window handle.
+	 */
+	public static String getProcessFilePath(final HWND hwnd) {
+		return getInstance().getProcessFilePath(hwnd);
+	}
+
+	/**
+	 * Requests the location and size of the window associated with the
+	 * specified window handle.
+	 *
+	 * @param hwnd
+	 *            The concerning window handle.
+	 * @return The location and size of the window.
+	 */
+	public static Rectangle getWindowLocationAndSize(final HWND hwnd) {
+		return getInstance().getWindowLocationAndSize(hwnd);
+	}
 }